requirements.txt
PyFunctional==1.3.0
